requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart  
pydantic[email]
pydantic-settings
celery==5.4.0
redis==5.2.0
